Noah + Copilot review
