+++ tools/synchro.py
@@ -190,6 +190,7 @@ def async_only_test(f: Path) -> bool:
         "test_async_loop_safety.py",
         "test_async_contextvars_reset.py",
         "test_async_loop_unblocked.py",
+        "test_async_network_layer.py",
     ]
